+++ lib/action_controller/serialization.rb
@@ -47,6 +47,7 @@ module ActionController
 
     [:_render_option_json, :_render_with_renderer_json].each do |renderer_method|
       define_method renderer_method do |resource, options|
+        options.fetch(:context) { options[:context] = request }
         serializable_resource = get_serializer(resource, options)
         super(serializable_resource, options)
       end