+++ lib/active_model/serializer/configuration.rb
@@ -7,6 +7,7 @@ module ActiveModel
       included do |base|
         base.config.array_serializer = ActiveModel::Serializer::ArraySerializer
         base.config.adapter = :flatten_json
+        base.config.jsonapi_resource_type = :plural
       end
     end
   end