+++ test/generators/serializer_generator_test.rb
@@ -46,7 +46,11 @@ class SerializerGeneratorTest < Rails::Generators::TestCase
   def test_with_no_attributes_does_not_add_extra_space
     run_generator ["account"]
     assert_file "app/serializers/account_serializer.rb" do |content|
-      assert_no_match /\n\nend/, content
+      if RUBY_PLATFORM =~ /mingw/
+        assert_no_match(/\r\n\r\nend/, content)
+      else
+        assert_no_match(/\n\nend/, content)
+      end
     end
   end
 end